requirements.txt
-i https://pypi.org/simple
dj-database-url==0.5.0
django==2.2.7
psycopg2==2.8.4
python-dotenv==0.10.3
pytz==2019.3
sqlparse==0.3.0
